backend/pyproject.toml
[tool.poetry]
name = "note"
version = "0.1.0"
description = "note app backend"
authors = ["blackboxsql <pythonboy007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.11"
graphene-django = "^2.15.0"
django-graphql-auth = "^0.3.16"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
django-graphql-jwt = "0.3.0"
graphene-django-cud = "^0.9.1"
django-seed = "^0.3.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
